Maintenance anti-rush mob
Adds a new mob: the genetic ambusher

With genetics soon getting a rework, rogue scientists have started to play god.

Mechanically, the purpose of this mob is to discourage rushing. It's a fairly weak mob, if you're paying attention. However, running through at mach speeds thinking a flash will be more than enough is likely to get you dropped on by a *moving* shadow. The shadow itself is also a mob, and thus can (and is meant to) be shot at.

Current faction is roach to prevent giving even more utility to spider chmants, despite the mob being in "giant spider".